v2.3
UT and OB done
Most validations done

To do
Spoiled

+++ HRISV2TestReact/app/Http/Controllers/UTController.php
@@ -17,11 +17,25 @@ class UTController extends Controller
     public function index()
     {
         $currentUser = Auth::user()->emp_no;
-        $utList = UTModel::with(['user', 'status'])->where('emp_no', $currentUser)->get();
+        $utList = UTModel::with(['user', 'status'])
+            ->where('emp_no', $currentUser)
+            ->whereIn('ut_status_id', [1, 2, 3])
+            ->get()
+            ->map(function ($ut) {
+                $approver = User::where('emp_no', $ut->approved_by)->first();
+                $ut->approver_name = $approver ? $approver->name : null;
+                return $ut;
+            });
+        $spoiledUTList = UTModel::with(['user', 'status'])
+            ->where('emp_no', $currentUser)
+            ->where('ut_status_id', 4)
+            ->get();
         return Inertia::render('UT_Module/ut_entry', [
-            'UTList' => $utList
+            'UTList' => $utList,
+            'spoiledUTList' => $spoiledUTList,
         ]);
     }
+
 
     public function UTEntry(Request $request): RedirectResponse
     {
@@ -53,9 +67,21 @@ class UTController extends Controller
         ]);
         return redirect('/UT_Module/ut_entry')->with('success', 'dz');
     }
+
+    public function viewUTRequest($id)
+    {
+        $viewUTRequest = UTModel::findorfail($id);
+        /*  dd("puke", $id, $viewUTRequest); */
+        return Inertia::render('UT_Module/ut_entry', [
+            'viewUTRequest' => $viewUTRequest,
+        ]);
+    }
+
     public function editUTRequest(Request $request)
     {
-        $ut = UTModel::where('ut_no', $request->ut_no)->first();
+        $ut = UTModel::where('ut_no', $request->ut_no) //Can only edit Pending
+            ->where('ut_status_id', 1)
+            ->first();
         $user = Auth::user()->emp_no;
         $request->validate([
             'ut_date' => 'required',
@@ -75,19 +101,27 @@ class UTController extends Controller
         return redirect()->intended('/UT_Module/ut_entry');
     }
 
-    public function viewUTRequest($id)
+    public function spoilUTRequest(Request $request, $id)
     {
-        $viewUTRequest = UTModel::findorfail($id);
-        /*  dd("puke", $id, $viewUTRequest); */
-        return Inertia::render('UT_Module/ut_entry', [
-            'viewUTRequest' => $viewUTRequest,
+        $spoilUTRequest = UTModel::where('id', $id)->where('ut_status_id', 1)->first(); //only pending can be deleted
+        $validated = $request->validate([
+            'ut_status_id' => 'required|in:4',
         ]);
+        $currentUser = Auth::user()->emp_no;
+        $spoilUTRequest->update([
+            'ut_status_id' => $validated['ut_status_id'],
+            'updated_by' => $currentUser,
+            'updated_date' => Carbon::now(),
+        ]);
+        return redirect()->intended('/UT_Module/ut_entry');
     }
-
     public function deleteUTRequest($id)
     {
-        $deleteUTRequest = UTModel::findorfail($id);
-        $deleteUTRequest->delete();
+        $deleteUTRequest = UTModel::where('id', $id)->where('ut_status_id', 1)->first(); //only pending can be deleted
+
+        if ($deleteUTRequest) {
+            $deleteUTRequest->delete();
+        }
 
         return redirect()->intended('/UT_Module/ut_entry');
     }
@@ -97,23 +131,34 @@ class UTController extends Controller
     public function UTApprList()
     {
         $apprvID = Auth::user()->emp_no;
-        $appr_pendings = UTModel::with(['user', 'status'])->where(function ($query) use ($apprvID) {
+        $appr_pendings = UTModel::with(['user', 'status'])->where(function ($query) use ($apprvID) { //Can view requests regardless if 1st or 2nd appr
             $query->where('first_apprv_no', $apprvID)
                 ->orWhere('sec_apprv_no', $apprvID);
         })
             ->where(function ($query) {
                 $query->where('ut_status_id', '1');
             })
-            ->get();
+            ->get()
+            ->map(function ($ut) {
+                $approver = User::where('emp_no', $ut->approved_by)->first();
+                $ut->approver_name = $approver ? $approver->name : null; //match emp_no approved by to user emp_no get name
+
+                return $ut;
+            });
         $appr_updated = UTModel::with(['user', 'status'])->where(function ($query) use ($apprvID) {
             $query->where('first_apprv_no', $apprvID)
                 ->orWhere('sec_apprv_no', $apprvID);
         })
             ->where(function ($query) {
-                $query->where('ut_status_id', '2')
-                    ->orWhere('ut_status_id', '3');
+                $query->where('ut_status_id', '2') //approved
+                    ->orWhere('ut_status_id', '3'); //disappored
             })
-            ->get();
+            ->get()
+            ->map(function ($ut) {
+                $approver = User::where('emp_no', $ut->approved_by)->first();
+                $ut->approver_name = $approver ? $approver->name : null;
+                return $ut;
+            });
         return Inertia::render('UT_Module/ut_appr_list', [
             'UTPendingList' => $appr_pendings,
             'UTUpdatedList' => $appr_updated,
@@ -121,31 +166,7 @@ class UTController extends Controller
     }
 
 
-    public function updateUTRequest(Request $request, $id)
-    {
-        $validated = $request->validate([
-            'ut_status_id' => 'required|in:2,3', // Validate that it is either 2 or 3
-        ]);
-        $currentUser = Auth::user()->emp_no;
-        $ut_status_id = $validated['ut_status_id'];
-
-
-        $remarks = $request->input('remarks');
-        $selectedUT = UTModel::findOrFail($id);
-        $selectedUT->update([
-            'ut_status_id' => $ut_status_id,
-            'remarks' => $remarks,
-            'approved_by' => $currentUser,
-            'approved_date' => Carbon::now(),
-            'updated_by' => $currentUser,
-            'updated_date' => Carbon::now(),
-
-        ]);
-        $selectedUT->refresh();
-        return redirect()->intended('/UT_Module/ut_appr_list');
-    }
-
-    public function updateAll(Request $request)
+    public function updateAll(Request $request) //Batch Update
     {
         $currentUser = Auth::user()->emp_no;
         $validated = $request->validate(['action' => 'required',]);
@@ -170,6 +191,28 @@ class UTController extends Controller
         return redirect()->intended('/UT_Module/ut_appr_list');
     }
 
+    public function updateUTRequest(Request $request, $id) //Quickedit
+    {
+        $validated = $request->validate([
+            'ut_status_id' => 'required|in:2,3', // Validate that it is either 2 or 3
+        ]);
+        $currentUser = Auth::user()->emp_no;
+        $ut_status_id = $validated['ut_status_id'];
+
+
+        $selectedUT = UTModel::findOrFail($id);
+        $selectedUT->update([
+            'ut_status_id' => $ut_status_id,
+            'approved_by' => $currentUser,
+            'approved_date' => Carbon::now(),
+            'updated_by' => $currentUser,
+            'updated_date' => Carbon::now(),
+
+        ]);
+        $selectedUT->refresh();
+        return redirect()->intended('/UT_Module/ut_appr_list');
+    }
+
     public function viewUTPendingRequest($id)
     {
         $viewUTPendingRequest = UTModel::findorfail($id);
@@ -178,15 +221,16 @@ class UTController extends Controller
             'viewUTPendingRequest' => $viewUTPendingRequest,
         ]);
     }
-    public function editUTApprRequest(Request $request)
+    public function editUTApprRequest(Request $request) //MODAL Edit
     {
-        $ut = UTModel::where('ut_no', $request->ut_no)->first();
+        $ut = UTModel::where('ut_no', $request->ut_no)
+            ->where('ut_status_id', 1)
+            ->first();
         $validated = $request->validate([
             'ut_status_id' => 'required|in:2,3',
             'remarks' => 'nullable|string' // Validate that it is either 2 or 3
         ]);
         $currentUser = Auth::user()->emp_no;
-        $statusname = $validated['ut_status_id'] == 2 ? 'Approved' : 'Rejected';
 
         $ut->update([
             'ut_status_id' => $validated['ut_status_id'],
@@ -196,7 +240,6 @@ class UTController extends Controller
             'updated_by' => $currentUser,
             'updated_date' => Carbon::now(),
         ]);
-        /* return response()->json(['message' => 'UT request updated successfully!', 'ut' => $ut]); */
         return redirect()->intended('/UT_Module/ut_appr_list');
     }
 
@@ -205,7 +248,18 @@ class UTController extends Controller
     {
         $apprvID = Auth::user()->emp_no;
 
-        $appr_list = UTModel::with(['user', 'status'])->where('first_apprv_no', $apprvID)->orWhere('sec_apprv_no', $apprvID)->get();
+        $appr_list = UTModel::with(['user', 'status'])
+            ->where('first_apprv_no', $apprvID)
+            ->orWhere('sec_apprv_no', $apprvID)
+            ->get()->map(function ($ut) {
+                $approver = User::where('emp_no', $ut->approved_by)->first();
+                $ut->approver_name = $approver ? $approver->name : null; //matahthacmah
+                $first_appr = User::where('emp_no', $ut->first_apprv_no)->first();
+                $ut->first_approver = $first_appr ? $first_appr->name : null; //match first_apprv_no to emp_no
+                $sec_appr = User::where('emp_no', $ut->sec_apprv_no)->first();
+                $ut->sec_approver = $sec_appr ? $sec_appr->name : null; //match sec_apprv_no to emp_no
+                return $ut;
+            });
         return Inertia::render('UT_Module/ut_reports_list', [
             'UTReportsList' => $appr_list,
         ]);
